requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asgi-csrf==0.11
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.3.0
autoflake==2.3.1
backports.asyncio.runner==1.2.0
bcrypt==4.3.0
binance==0.3.51
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-default-group==1.2.4
colorama==0.4.6
contourpy==1.3.2
coverage==7.9.2
cryptography==45.0.4
cycler==0.12.1
datasette==0.65.1
dateparser==1.2.2
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.14
flexcache==0.3
flexparser==0.4
fonttools==4.60.1
freezegun==1.5.3
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
hupper==1.12.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
janus==2.0.0
Jinja2==3.1.6
jiter==0.11.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.7
mergedeep==1.3.4
multidict==6.6.0
numpy==1.26.4
openai==2.6.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
pandas-ta==0.3.14b0
passlib==1.7.4
pillow==11.2.1
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
pytest-mock==3.14.1
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
starlette==0.46.2
SuperClaude==3.0.0.2
tenacity==8.5.0
tomli==2.2.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.35.0
watchfiles==1.1.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
